refactor: products class size optimization; add new class, add helpers component
